requirements.txt
beautifulsoup4==4.7.1
boltons==18.0.1
grpcio==1.18.0
html5lib==1.0.1
inflection==0.3.1
interruptingcow==0.8
jsbeautifier==1.8.9
libarchive-c==2.8
lxml==4.3.0
git+https://github.com/aaronst/macholibre.git       # v1.0 as of this freeze
olefile==0.46
oletools==0.53.1
pdfminer.six==20181108
pebble==4.3.10
pefile==2018.8.8
pgpdump3==1.5.2
protobuf==3.6.1
pyelftools==0.25
pyjsparser==2.5.2
pylzma==0.5.0
pygments==2.3.1
pyopenssl==18.0.0
python-docx==0.8.10
git+https://github.com/jshlbrd/python-entropy.git   # v0.11 as of this freeze (package installed as 'entropy')
python-magic==0.4.15
pyyaml>=4.2b1
rarfile==3.0
requests==2.21.0
rpmfile==0.1.4
ssdeep==3.3
tnefparse==1.3.0
